Consolidate and cleanup close_fd functionality
Since it gets used multiple times, put the code in one place. Provide a fast and slow path depending on available support
